Move plug-in loading logic to PluginLoader
